pull

+++ app/src/main/java/com/besome/sketch/adapters/ProjectsAdapter.java
@@ -3,23 +3,21 @@ package com.besome.sketch.adapters;
 import android.app.Activity;
 import android.content.Intent;
 import android.net.Uri;
-import android.os.Build;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.annotation.DrawableRes;
 import androidx.annotation.NonNull;
 import androidx.core.content.FileProvider;
 import androidx.recyclerview.widget.DiffUtil;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.besome.sketch.export.ExportProjectActivity;
-import pro.sketchware.activities.main.fragments.projects.ProjectsFragment;
+import com.besome.sketch.lib.ui.LoadingDialog;
 import com.besome.sketch.projects.MyProjectSettingActivity;
 import com.google.android.material.bottomsheet.BottomSheetDialog;
-import pro.sketchware.R;
-import pro.sketchware.databinding.BottomSheetProjectOptionsBinding;
-import pro.sketchware.databinding.MyprojectsItemBinding;
+import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -27,15 +25,20 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 
-import a.a.a.ZA;
-import a.a.a.aB;
 import a.a.a.lC;
 import a.a.a.mB;
 import a.a.a.wq;
 import a.a.a.yB;
+import extensions.anbui.daydream.activity.DayDreamBackupTool;
+import extensions.anbui.daydream.activity.DayDreamProjectTool;
+import extensions.anbui.daydream.configs.Configs;
 import mod.hey.studios.project.ProjectSettingsDialog;
 import mod.hey.studios.project.backup.BackupRestoreManager;
 import mod.hey.studios.util.Helper;
+import pro.sketchware.R;
+import pro.sketchware.activities.main.fragments.projects.ProjectsFragment;
+import pro.sketchware.databinding.BottomSheetProjectOptionsBinding;
+import pro.sketchware.databinding.MyprojectsItemBinding;
 
 public class ProjectsAdapter extends RecyclerView.Adapter<ProjectsAdapter.ProjectViewHolder> {
     private final ProjectsFragment projectsFragment;
@@ -45,7 +48,7 @@ public class ProjectsAdapter extends RecyclerView.Adapter<ProjectsAdapter.Projec
 
     public ProjectsAdapter(ProjectsFragment projectsFragment, List<HashMap<String, Object>> allProjects) {
         this.projectsFragment = projectsFragment;
-        this.activity = projectsFragment.requireActivity();
+        activity = projectsFragment.requireActivity();
         this.allProjects = allProjects;
     }
 
@@ -97,6 +100,7 @@ public class ProjectsAdapter extends RecyclerView.Adapter<ProjectsAdapter.Projec
         }, true);
         shownProjects = newProjects;
         result.dispatchUpdatesTo(this);
+        notifyDataSetChanged();
     }
 
     @Override
@@ -114,8 +118,21 @@ public class ProjectsAdapter extends RecyclerView.Adapter<ProjectsAdapter.Projec
         return false;
     }
 
+    @DrawableRes
+    public static <T> int getShapedBackgroundForList(List<T> list, int position) {
+        if (list.size() == 1) {
+            return R.drawable.project_item_shape_alone;
+        } else if (position == 0) {
+            return R.drawable.project_item_shape_top;
+        } else if (position == list.size() - 1) {
+            return R.drawable.project_item_shape_bottom;
+        } else {
+            return R.drawable.project_item_shape_middle;
+        }
+    }
     @Override
     public void onBindViewHolder(@NonNull ProjectViewHolder holder, int position) {
+        holder.itemView.setBackgroundResource(getShapedBackgroundForList(shownProjects, position));
         HashMap<String, Object> projectMap = shownProjects.get(position);
         String scId = yB.c(projectMap, "sc_id");
 
@@ -135,12 +152,8 @@ public class ProjectsAdapter extends RecyclerView.Adapter<ProjectsAdapter.Projec
         if (yB.a(projectMap, "custom_icon")) {
             String iconFolder = wq.e() + File.separator + scId;
             Uri uri;
-            if (Build.VERSION.SDK_INT >= 24) {
-                String providerPath = activity.getPackageName() + ".provider";
-                uri = FileProvider.getUriForFile(activity, providerPath, new File(iconFolder, "icon.png"));
-            } else {
-                uri = Uri.fromFile(new File(iconFolder, "icon.png"));
-            }
+            String providerPath = activity.getPackageName() + ".provider";
+            uri = FileProvider.getUriForFile(activity, providerPath, new File(iconFolder, "icon.png"));
             holder.binding.imgIcon.setImageURI(uri);
         }
 
@@ -181,17 +194,8 @@ public class ProjectsAdapter extends RecyclerView.Adapter<ProjectsAdapter.Projec
         return new ProjectViewHolder(binding);
     }
 
-    static class ProjectViewHolder extends RecyclerView.ViewHolder {
-        final MyprojectsItemBinding binding;
-
-        ProjectViewHolder(MyprojectsItemBinding binding) {
-            super(binding.getRoot());
-            this.binding = binding;
-        }
-    }
-
     private void deleteProject(HashMap<String, Object> projectMap, int position) {
-        ZA progressDialog = new ZA(activity);
+        LoadingDialog progressDialog = new LoadingDialog(activity);
         progressDialog.show();
 
         String scId = yB.c(projectMap, "sc_id");
@@ -200,7 +204,7 @@ public class ProjectsAdapter extends RecyclerView.Adapter<ProjectsAdapter.Projec
             activity.runOnUiThread(() -> {
                 progressDialog.dismiss();
                 shownProjects.remove(position);
-                notifyItemRemoved(position);
+                notifyDataSetChanged();
                 allProjects.remove(projectMap);
             });
         }).start();
@@ -257,19 +261,35 @@ public class ProjectsAdapter extends RecyclerView.Adapter<ProjectsAdapter.Projec
             projectOptionsBSD.dismiss();
         });
 
+        binding.projectDaydream.setOnClickListener(v -> {
+            Intent intent = new Intent(activity, DayDreamProjectTool.class);
+            intent.putExtra("sc_id", yB.c(projectMap, "sc_id"));
+            activity.startActivity(intent);
+            projectOptionsBSD.dismiss();
+        });
+
         binding.projectDelete.setOnClickListener(v -> {
             projectOptionsBSD.dismiss();
-            aB dialog = new aB(activity);
-            dialog.a(R.drawable.icon_delete);
-            dialog.b(Helper.getResString(R.string.delete_project_dialog_title));
-            dialog.a(Helper.getResString(R.string.delete_project_dialog_message).replace("%1$s", yB.c(projectMap, "my_app_name")));
-            dialog.b(Helper.getResString(R.string.common_word_delete), v1 -> {
+            MaterialAlertDialogBuilder dialog = new MaterialAlertDialogBuilder(activity);
+            dialog.setIcon(R.drawable.icon_delete);
+            dialog.setTitle(Helper.getResString(R.string.delete_project_dialog_title));
+            dialog.setMessage(Helper.getResString(R.string.delete_project_dialog_message).replace("%1$s", yB.c(projectMap, "my_app_name")));
+            dialog.setPositiveButton(Helper.getResString(R.string.common_word_delete), (v1, which) -> {
                 deleteProject(projectMap, position);
-                dialog.dismiss();
+                v1.dismiss();
             });
-            dialog.a(Helper.getResString(R.string.common_word_cancel), Helper.getDialogDismissListener(dialog));
+            dialog.setNegativeButton(Helper.getResString(R.string.common_word_cancel), null);
             dialog.show();
         });
         projectOptionsBSD.show();
+    }
+
+    static class ProjectViewHolder extends RecyclerView.ViewHolder {
+        final MyprojectsItemBinding binding;
+
+        ProjectViewHolder(MyprojectsItemBinding binding) {
+            super(binding.getRoot());
+            this.binding = binding;
+        }
     }
 }
