+++ app/src/main/java/pro/sketchware/activities/editor/component/ManageCustomComponentActivity.java
@@ -4,7 +4,6 @@ import static pro.sketchware.utility.GsonUtils.getGson;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.os.Environment;
 import android.os.Parcelable;
 import android.util.SparseBooleanArray;
 import android.view.LayoutInflater;
@@ -20,17 +19,16 @@ import android.widget.ListView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
+import androidx.core.graphics.Insets;
+import androidx.core.view.ViewCompat;
+import androidx.core.view.WindowInsetsCompat;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.activity.EdgeToEdge;
 
+import com.besome.sketch.lib.base.BaseAppCompatActivity;
 import com.besome.sketch.lib.base.CollapsibleViewHolder;
 import com.besome.sketch.lib.ui.CollapsibleButton;
-import com.besome.sketch.lib.base.BaseAppCompatActivity;
-import com.github.angads25.filepicker.model.DialogConfigs;
-import com.github.angads25.filepicker.model.DialogProperties;
-import com.github.angads25.filepicker.view.FilePickerDialog;
 import com.google.android.material.appbar.MaterialToolbar;
-import com.google.gson.Gson;
+import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -40,33 +38,30 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import a.a.a.aB;
 import a.a.a.wq;
-
+import dev.pranav.filepicker.FilePickerCallback;
+import dev.pranav.filepicker.FilePickerDialogFragment;
+import dev.pranav.filepicker.FilePickerOptions;
 import mod.hey.studios.util.Helper;
 import mod.hilal.saif.components.ComponentsHandler;
 import mod.jbk.util.OldResourceIdMapper;
-
 import pro.sketchware.R;
 import pro.sketchware.utility.FileUtil;
 import pro.sketchware.utility.SketchwareUtil;
 
 public class ManageCustomComponentActivity extends BaseAppCompatActivity {
 
-    private List<HashMap<String, Object>> componentsList = new ArrayList<>();
-
     private static final String COMPONENT_EXPORT_DIR = wq.getExtraDataExport() + "/components/";
     private static final String COMPONENT_DIR = wq.getCustomComponent();
-
+    private List<HashMap<String, Object>> componentsList = new ArrayList<>();
     private TextView tv_guide;
     private RecyclerView componentView;
 
     @Override
     public void onCreate(Bundle _savedInstanceState) {
-        EdgeToEdge.enable(this);
+        enableEdgeToEdgeNoContrast();
         super.onCreate(_savedInstanceState);
         setContentView(R.layout.manage_custom_component);
-        handleInsetts(findViewById(R.id.root));
         init();
     }
 
@@ -82,6 +77,49 @@ public class ManageCustomComponentActivity extends BaseAppCompatActivity {
 
         findViewById(R.id.fab).setOnClickListener(_view ->
                 startActivity(new Intent(getApplicationContext(), AddCustomComponentActivity.class)));
+
+        {
+            View view = findViewById(R.id.content);
+            int left = view.getPaddingLeft();
+            int top = view.getPaddingTop();
+            int right = view.getPaddingRight();
+            int bottom = view.getPaddingBottom();
+
+            ViewCompat.setOnApplyWindowInsetsListener(view, (v, i) -> {
+                Insets insets = i.getInsets(WindowInsetsCompat.Type.systemBars() | WindowInsetsCompat.Type.ime() | WindowInsetsCompat.Type.displayCutout());
+                v.setPadding(left + insets.left, top, right + insets.right, bottom + insets.bottom);
+                return i;
+            });
+        }
+
+        {
+            View view = findViewById(R.id.app_bar_layout);
+            int left = view.getPaddingLeft();
+            int top = view.getPaddingTop();
+            int right = view.getPaddingRight();
+            int bottom = view.getPaddingBottom();
+
+            ViewCompat.setOnApplyWindowInsetsListener(view, (v, i) -> {
+                Insets insets = i.getInsets(WindowInsetsCompat.Type.systemBars() | WindowInsetsCompat.Type.displayCutout());
+                v.setPadding(left + insets.left, top + insets.top, right + insets.right, bottom + insets.bottom);
+                return i;
+            });
+        }
+
+        {
+            View view = findViewById(R.id.fab);
+            ViewGroup.MarginLayoutParams lp = (ViewGroup.MarginLayoutParams) view.getLayoutParams();
+            int end = lp.getMarginEnd();
+            int bottom = lp.bottomMargin;
+
+            ViewCompat.setOnApplyWindowInsetsListener(view, (v, i) -> {
+                Insets insets = i.getInsets(WindowInsetsCompat.Type.systemBars() | WindowInsetsCompat.Type.displayCutout());
+                lp.setMarginEnd(end + insets.right);
+                lp.bottomMargin = bottom + insets.bottom;
+                v.setLayoutParams(lp);
+                return i;
+            });
+        }
     }
 
     @Override
@@ -120,7 +158,7 @@ public class ManageCustomComponentActivity extends BaseAppCompatActivity {
         }
     }
 
-    private void readComponents(final String _path) {
+    private void readComponents(String _path) {
         componentsList = getGson().fromJson(FileUtil.readFile(_path), Helper.TYPE_MAP_LIST);
         if (componentsList != null && !componentsList.isEmpty()) {
             ComponentsAdapter adapter = new ComponentsAdapter(componentsList);
@@ -137,21 +175,20 @@ public class ManageCustomComponentActivity extends BaseAppCompatActivity {
     }
 
     private void showFilePickerDialog() {
-        DialogProperties properties = new DialogProperties();
+        FilePickerOptions options = new FilePickerOptions();
+        options.setTitle("Select .json selector file");
+        options.setExtensions(new String[]{"json"});
 
-        properties.selection_mode = DialogConfigs.SINGLE_MODE;
-        properties.selection_type = DialogConfigs.FILE_SELECT;
-        properties.root = Environment.getExternalStorageDirectory();
-        properties.error_dir = Environment.getExternalStorageDirectory();
-        properties.offset = Environment.getExternalStorageDirectory();
-        properties.extensions = new String[]{"json"};
+        FilePickerCallback callback = new FilePickerCallback() {
+            @Override
+            public void onFileSelected(File file) {
+                selectComponentToImport(file.getAbsolutePath());
+            }
+        };
 
-        FilePickerDialog pickerDialog = new FilePickerDialog(this, properties, R.style.RoundedCornersDialog);
+        FilePickerDialogFragment pickerDialog = new FilePickerDialogFragment(options, callback);
 
-        pickerDialog.setTitle("Select .json selector file");
-        pickerDialog.setDialogSelectionListener(selections -> selectComponentToImport(selections[0]));
-
-        pickerDialog.show();
+        pickerDialog.show(getSupportFragmentManager(), "filePickerDialog");
     }
 
     private void selectComponentToImport(String path) {
@@ -166,8 +203,8 @@ public class ManageCustomComponentActivity extends BaseAppCompatActivity {
                 .map(component -> (String) component.get("name"))
                 .collect(Collectors.toList());
         if (componentNames.size() > 1) {
-            var dialog = new aB(this);
-            dialog.b(Helper.getResString(R.string.logic_editor_title_select_component));
+            MaterialAlertDialogBuilder dialog = new MaterialAlertDialogBuilder(this);
+            dialog.setTitle(Helper.getResString(R.string.logic_editor_title_select_component));
             ArrayList<Integer> selectedPositions = new ArrayList<>();
             var listView = new ListView(this);
             ArrayAdapter<String> arrayAdapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_multiple_choice, componentNames);
@@ -184,8 +221,8 @@ public class ManageCustomComponentActivity extends BaseAppCompatActivity {
                     selectedPositions.remove(Integer.valueOf(position));
                 }
             });
-            dialog.a(listView);
-            dialog.b(Helper.getResString(R.string.common_word_import), v -> {
+            dialog.setView(listView);
+            dialog.setPositiveButton(Helper.getResString(R.string.common_word_import), (v, which) -> {
                 for (int position : selectedPositions) {
                     var component = components.get(position);
                     if (position != -1 && ComponentsHandler.isValidComponent(component)) {
@@ -196,9 +233,9 @@ public class ManageCustomComponentActivity extends BaseAppCompatActivity {
                 }
                 FileUtil.writeFile(COMPONENT_DIR, getGson().toJson(componentsList));
                 readSettings();
-                dialog.dismiss();
+                v.dismiss();
             });
-            dialog.a(Helper.getResString(R.string.common_word_cancel), Helper.getDialogDismissListener(dialog));
+            dialog.setNegativeButton(Helper.getResString(R.string.common_word_cancel), null);
             dialog.show();
         } else {
             var component = components.get(0);
@@ -212,25 +249,25 @@ public class ManageCustomComponentActivity extends BaseAppCompatActivity {
         }
     }
 
-    private void save(final HashMap<String, Object> _item) {
+    private void save(HashMap<String, Object> _item) {
         componentsList.remove(_item);
         FileUtil.writeFile(COMPONENT_DIR, getGson().toJson(componentsList));
     }
 
     private void export(int position) {
         String componentName = componentsList.get(position).get("name").toString();
-        var dialog = new aB(this);
-        dialog.b(Helper.getResString(R.string.common_word_export));
-        dialog.a(Helper.getResString(R.string.developer_tools_component_message_export, componentName));
-        dialog.a(R.drawable.export_96);
-        dialog.b(Helper.getResString(R.string.common_word_yes), v -> {
+        MaterialAlertDialogBuilder dialog = new MaterialAlertDialogBuilder(this);
+        dialog.setTitle(Helper.getResString(R.string.common_word_export));
+        dialog.setMessage(Helper.getResString(R.string.developer_tools_component_message_export, componentName));
+        dialog.setIcon(R.drawable.export_96);
+        dialog.setPositiveButton(Helper.getResString(R.string.common_word_yes), (v, which) -> {
             String fileName = componentName + ".json";
             String filePath = new File(COMPONENT_EXPORT_DIR, fileName).getAbsolutePath();
             FileUtil.writeFile(filePath, getGson().toJson(List.of(componentsList.get(position))));
             SketchwareUtil.toast(Helper.getResString(R.string.developer_tools_component_success_message_export, filePath));
-            dialog.dismiss();
+            v.dismiss();
         });
-        dialog.a(Helper.getResString(R.string.common_word_cancel), Helper.getDialogDismissListener(dialog));
+        dialog.setNegativeButton(Helper.getResString(R.string.common_word_cancel), null);
         dialog.show();
     }
 
@@ -240,9 +277,9 @@ public class ManageCustomComponentActivity extends BaseAppCompatActivity {
         private final List<Boolean> confirmation;
 
         public ComponentsAdapter(List<HashMap<String, Object>> itemList) {
-            this.components = itemList;
-            this.collapse = new ArrayList<>(Collections.nCopies(itemList.size(), true));
-            this.confirmation = new ArrayList<>(Collections.nCopies(itemList.size(), false));
+            components = itemList;
+            collapse = new ArrayList<>(Collections.nCopies(itemList.size(), true));
+            confirmation = new ArrayList<>(Collections.nCopies(itemList.size(), false));
         }
 
         @NonNull
@@ -346,7 +383,7 @@ public class ManageCustomComponentActivity extends BaseAppCompatActivity {
                 onDoneInitializingViews();
                 root.setOnClickListener(v -> {
                     Intent intent = new Intent(getApplicationContext(), AddCustomComponentActivity.class);
-                    intent.putExtra("pos", (int) getLayoutPosition());
+                    intent.putExtra("pos", getLayoutPosition());
                     startActivity(intent);
                 });
                 setOnClickCollapseConfig(v -> v != root);
